+++ spec/lib/tasks/couchdb_rake_spec.rb
@@ -1,0 +1,13 @@
+require 'spec_helper'
+
+describe "couchdb:histories:delete" do
+  include_context "rake"
+
+  let(:output) { "No CouchDB history documents since #{Date.today} to delete.\n" }
+
+  its(:prerequisites) { should include("environment") }
+
+  it "should run the rake task" do
+    capture_stdout { subject.invoke }.should eq(output)
+  end
+end